move exceptions out of http_client, retry all errors #126
